Massive Redo
Changed the setup a great deal to accommodate larger maps and a
multiview system based off a single map. A lot of it is pretty abstract
at the moment, so let me know if you need any explanation. The redo is
far from done, so the Screen and one test are bombing. Seriously, if
anybody can find out why that one Map test is bombing, that would make
me a happy panda.

+++ src/cs2114/tiletraveler/Location.java
@@ -9,8 +9,8 @@ package cs2114.tiletraveler;
  */
 public class Location
 {
-    private float xCoor;
-    private float yCoor;
+    private int xCoor;
+    private int yCoor;
 
     // ----------------------------------------------------------
     /**
@@ -18,7 +18,7 @@ public class Location
      * @param x x coordinate of location
      * @param y y coordinate of location
      */
-    public Location(float x, float y)
+    public Location(int x, int y)
     {
         xCoor = x;
         yCoor = y;
@@ -29,7 +29,7 @@ public class Location
      * Get the x of location
      * @return xCoor x coordinate of location
      */
-    public float x()
+    public int x()
     {
         return xCoor;
     }
@@ -39,7 +39,7 @@ public class Location
      * Get the y of location
      * @return yCoor y coordinate of location
      */
-    public float y()
+    public int y()
     {
         return yCoor;
     }
